+++ formula/src/main/java/com/instacart/formula/internal/OverrideKeyAction.kt
@@ -1,0 +1,12 @@
+package com.instacart.formula.internal
+
+import com.instacart.formula.Action
+import com.instacart.formula.DelegateAction
+
+internal class OverrideKeyAction<Event>(
+    val key: Any,
+    delegateAction: Action<Event>,
+) : DelegateAction<Event>(delegateAction) {
+
+    override fun key(): Any = key
+}